requirements.txt
mpmath==1.2.1
numpy==1.21.4
Pillow==8.4.0
sympy==1.9
